==> disciplinaryDatasetPreparation/0.select_journals_fromScimago.ipynb <==
import pandas as pd
import collections

# PREPARE JOURNALS

infile='Journals.nt'

def nameline(text):
    textclean=text.replace(';',',').replace('<https://makg.org/entity/','').replace('> <http://xmlns.com/foaf/0.1/name> ',';').replace('^^<http://www.w3.org/2001/XMLSchema#string> .','')
    return textclean.replace('"','').rstrip().split(';')

def issnline(text):
    textclean=text.replace('<https://makg.org/entity/','').replace('> <http://id.loc.gov/vocabulary/identifiers/issn>',';').replace('^^<http://www.w3.org/2001/XMLSchema#string> .','')
    return textclean.replace('"','').rstrip().split('; ')

name={}
issn={}


for line in open(infile):
    try:
        if 'issn>' in line:
            [u,v]=issnline(line)
            issn[u]=v.replace('-','')
        if 'name>' in line:
            [u,v]=nameline(line)
            name[u]=v
    except:
        print(line)


df=pd.DataFrame()
df['journalId']=name.keys()
df['name']=[name[i] for i in name.keys()]
df['issn']=[issn[i] if i in issn.keys() else 'NONE' for i in name.keys()]

df.to_csv('journals.csv',index=False)

# FILTER JOURNALS

df=pd.read_csv('journals.csv')

df.head()

infile='disciplinaryJournals/ Neuroscience.csv'
dfSMJ=pd.read_csv(infile, sep=';')
dfSMJ=dfSMJ[dfSMJ['Issn']!='-']

dfSMJ.head()

listIssn=(', ').join(dfSMJ['Issn']).split(', ')

dfNeuroscience=df[df['issn'].isin(listIssn)]

dfNeuroscience.to_csv('neuroscience_journals.csv',index=False)
