==> disciplinaryDatasetPreparation/1.selectDisciplinaryPapers.ipynb <==
import os
import pandas as pd
import re
import json

# 1. Extract the list from PaperJournals.nt

#your_path = '../../../../../../Volumes/Samsung_T5/MAGextracted29_5_2020/'
#print( os.listdir(your_path))

#the file PapersJournals has been prepared starting from papers.nt and selecting the journal lines, using unix command grep
infile='../PapersJournal.nt'

dfNeuroJ=pd.read_csv('../1_selectJournals/MAG_disciplinaryJournals/neuroscience_journals.csv')

neuroJ=list(dfNeuroJ['journalId'])

len(neuroJ)

centralString='> <http://ma-graph.org/property/appearsInJournal> '
entString='<http://ma-graph.org/entity/'

selectedPapers=[]
for line in open(infile):
    [p,j]=(line.rstrip().replace(centralString,';').replace(entString,'').replace('> .','').split(';'))
    if int(j) in neuroJ:
        selectedPapers.append(p)
    
   

    

outfile='paperLists/papers_neuroscience.txt'
ww=open(outfile,'w')
for el in selectedPapers:
    ww.write(el+'\n')
ww.close()

# 2. Build the filtered files for the selected papers

paperList=[]
for line in open('paperLists/papersList_neuroscience.txt'):
    paperList.append(line.rstrip())

your_path = '../../../../../../Volumes/TOSHIBA EXT/Simplified_files'
print( os.listdir(your_path))
infile=your_path+'/Papers_simplified.txt'

thisDf=pd.DataFrame()
chunksize = 10 ** 7
i=0
for chunk in pd.read_csv(infile, chunksize=chunksize, names=['id', 'feature','value'],sep=';'):
    thisDf=thisDf.append(chunk[chunk['id'].isin(paperList)])
    print (i)
    i=i+1

thisDf.to_csv('Papers_neuroscience_temp.txt', index=False)

thisDf=pd.read_csv('paperTemp/Papers_neuroscience_temp.txt')

len(thisDf)

thisDf

redDf=thisDf[thisDf['feature']!=' http://www.w3.org/1999/02/22-rdf-syntax-ns#type']
redDf=redDf[redDf['feature']!=' rank']
redDf=redDf[redDf['feature']!=' estimatedCitationCount']

feat=list(set(redDf['feature']))

feat

dictFeat={}
for f in feat:
    #dictFeat[f]={}
    df=redDf[redDf['feature']==f]
    dictFeat[f]=dict(zip(df['id'],df['value']))

dictFeat[' title']

ids=list(set(redDf['id']))

len(ids)

dictPapers=[]
for p in ids:
    thisdict={}
    thisdict['id']=p
    for f in dictFeat.keys():
        try:
            thisdict[f]=dictFeat[f][p]
        except:
            continue
    dictPapers.append(thisdict)
        
     

dictPapers[2]

with open('../neuroscience/Papers_neuroscience.json', 'w') as fout:
    json.dump(dictPapers , fout)
