==> disciplinaryDatasetPreparation/2.citationNetwork.ipynb <==
import os
import pandas as pd
import re
from collections import ChainMap
import json

with open('../neuroscience/Papers_neuroscience.json', "r") as read_file:
    papers = json.load(read_file)

your_path = '../../../../../../Volumes/TOSHIBA EXT/Simplified_files'
print( os.listdir(your_path))
infile=your_path+'/paperReferences_simplified.txt'

listPapers=[int(p['id']) for p in papers]

thisDf=pd.DataFrame()
chunksize = 10 ** 6
for chunk in pd.read_csv(infile, chunksize=chunksize, names=['citing', 'cited'],sep=','):
    thisDf=thisDf.append(chunk[chunk['citing'].isin(listPapers)])
    thisDf=thisDf.append(chunk[chunk['cited'].isin(listPapers)])

len(thisDf)

outfile='../neuroscience/citationNetwork.csv'
thisDf.to_csv(outfile)

allPapersCN=list(set(thisDf['citing'])|set(thisDf['cited']))

notInList=list(set(allPapersCN)-set(listPapers))

outFile='../neuroscience/temp/nextToSearch.txt'
ww=open(outFile,'w')
for el in notInList:
    ww.write(str(el)+'\n')
ww.close()

# ADD CITED and CITING papers

your_path = '../../../../../../Volumes/TOSHIBA EXT/Simplified_files'
print( os.listdir(your_path))
infile=your_path+'/Papers_simplified.txt'

newDf=pd.DataFrame()
chunksize = 5*10 ** 6
for chunk in pd.read_csv(infile, chunksize=chunksize, names=['id', 'feature','value'],sep=';'):
    newDf=newDf.append(chunk[chunk['id'].isin(notInList)])

redDf=newDf[newDf['feature']!=' http://www.w3.org/1999/02/22-rdf-syntax-ns#type']
redDf=newDf[newDf['feature']!=' rank']
redDf=newDf[newDf['feature']!=' estimatedCitationCount']

feat=list(set(redDf['feature']))

dictFeat={}
for f in feat:
    #dictFeat[f]={}
    df=redDf[redDf['feature']==f]
    dictFeat[f]=dict(zip(df['id'],df['value']))

ids=list(set(redDf['id']))

dictPapers=[]
for p in ids:
    thisdict={}
    thisdict['id']=p
    thisdict['queryLevel']=1
    for f in dictFeat.keys():
        try:
            thisdict[f]=dictFeat[f][p]
        except:
            continue
    dictPapers.append(thisdict)

dictPapers[0]

papers[0]

for p in papers:
    p['queryLevel']=0

allpapers=papers+dictPapers

with open('../neuroscience/Papers_neuroscience_all.json', 'w') as fout:
    json.dump(allpapers , fout)
