==> disciplinaryDatasetPreparation/3.selectAuthors.ipynb <==
import os
import pandas as pd
import re
from collections import ChainMap
import json

with open('../neuroscience/Papers_neuroscience_all.json', "r") as read_file:
    papers = json.load(read_file)

listPapers0=[p['id'] for p in papers if p['queryLevel']==0]
listPapers1=[p['id'] for p in papers if p['queryLevel']==1]

listPapers=listPapers0+listPapers1

your_path = '../../../../../../Volumes/TOSHIBA EXT/Simplified_files'
print( os.listdir(your_path))
infile=your_path+'/paperAuthors_simplified.txt'

selDf=pd.DataFrame()
chunksize = 10 ** 6
for chunk in pd.read_csv(infile, chunksize=chunksize, names=['paperId', 'authorId'],sep=';'):
    selDf=selDf.append(chunk[chunk['paperId'].isin(listPapers)])

d2=selDf.groupby('paperId')['authorId'].apply(list)

e=0
for p in papers:
    try:
        p['authors']=d2[int(p['id'])]
    except:
        e=e+1

with open('../neuroscience/Papers_neuroscience_all.json', 'w') as fout:
    json.dump(papers , fout)

# preparation author affiliation csv

papers[0]

ll=[el['authors'] for el in papers if 'authors' in el.keys()]
flat_list = [item for sublist in ll for item in sublist]
allAu=list(set(flat_list))

p0=[p for p in papers if p['queryLevel']==0]
p1=[p for p in papers if p['queryLevel']==1]

ll0=[el['authors'] for el in p0 if 'authors' in el.keys()]
ll1=[el['authors'] for el in p1 if 'authors' in el.keys()]

flat_list = [item for sublist in ll0 for item in sublist]
au0=set(flat_list)

flat_list = [item for sublist in ll1 for item in sublist]
au1=set(flat_list)

only1=list(au1-au0)

queryLevel={}
for el in list(au0):
    queryLevel[el]=0
for el in only1:
    queryLevel[el]=1

your_path = '../../../../../../Volumes/TOSHIBA EXT/Simplified_files'
infile=your_path+'/Authors_simplified.txt'

auDf=pd.DataFrame()
chunksize = 10 ** 6
for chunk in pd.read_csv(infile, chunksize=chunksize, names=['authorId', 'feature','value'],sep=';'):
    auDf=auDf.append(chunk[chunk['authorId'].isin(allAu)])

aff={}
name={}
affDf=auDf[auDf['feature']=='affiliation']
nameDf=auDf[auDf['feature']=='name']

for (u,v) in zip(affDf['authorId'],affDf['value']):
    aff[u]=v
    
for (u,v) in zip(nameDf['authorId'],nameDf['value']):
    name[u]=v

auAffDf=pd.DataFrame()
auAffDf['authorId']=name.keys()
auAffDf['name']=[name[i] for i in name.keys()]
auAffDf['affiliation']=[aff[i] if i in aff.keys() else 'NONE' for i in name.keys()]
auAffDf['queryLevel']=[queryLevel[i] for i in name.keys()]

auAffDf.to_csv('../neuroscience/authorId_name_affiliation_neuroscience.csv', index=False)
