==> disciplinaryDatasetPreparation/4.affiliation_identifiers.ipynb <==
import os
import pandas as pd
import re
import json

with open('../neuroscience/Papers_neuroscience_all.json', "r") as read_file:
    papers = json.load(read_file)

infile='../neuroscience/authorId_name_affiliation_neuroscience.csv'
df=pd.read_csv(infile)

affs=list(set(df[df['affiliation']!='NONE']['affiliation']))

your_path = '../../../../../../Volumes/TOSHIBA EXT/Simplified_files'
print( os.listdir(your_path))
infile=your_path+'/affiliations.json'

with open(infile, 'r') as read_file:
    allAff = json.load(read_file)

thisAff=[i for i in allAff if i['affiliation_id'] in affs]

len(thisAff)

with open('../neuroscience/Affiliations_neuroscience.json', 'w') as fout:
    json.dump(thisAff , fout)
