==> disciplinaryDatasetPreparation/5.addAbstracts.ipynb <==
import os
import pandas as pd
import re
import json

with open('../neuroscience/Papers_neuroscience_all.json', "r") as read_file:
    papers = json.load(read_file)

your_path = '../../../../../../Volumes/TOSHIBA EXT/Simplified_files'
print( os.listdir(your_path))
infile=your_path+'/paperAbstracts_simplified.txt'

listPId=[i['id'] for i in papers]

thisDf=pd.DataFrame()
chunksize = 10 ** 6
for chunk in pd.read_csv(infile, chunksize=chunksize, names=['paperId', 'abstract'],sep=';'):
    thisDf=thisDf.append(chunk[chunk['paperId'].isin(listPId)])

paperAbs=dict(zip(thisDf['paperId'], thisDf['abstract']))

len(papers)

for p in papers:
    if p['id'] in paperAbs.keys():
        p['Abstract']=paperAbs[p['id']]

with open('../sociology/Papers_sociology_all.json', 'w') as fout:
    json.dump(papers , fout)
